streamlit_app/requirements.txt
tensorflow
numpy
opencv-python
Pillow
ultralytics
streamlit
matplotlib
scikit-learn
